fix: ahci: cp110: add option to modify OOB parameters
Out of band (OOB) signals are used for handshake signaling
between a device and its host.
There are three OOB signals: COMRESET, COMINIT and COMWAKE.
This patch allows to modify the COMRESET and COMWAKE limits in order
to improve SATA electrical behavior for systems that
can't use the default values.

The new values should be given by the device-tree.
Binding document updated accordingly.

Change-Id: I4a17fba1c6fe786005669d31217f961a456fd4ff
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37097
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>