crypto: inside-secure: add eip197 basic driver
This patch adds a driver which performs a basic initialization
of the eip197 hardware including firmware, cache and PRNG support.

The PacketEngine-IP-197 (EIP-197) is IP for accelerating IPSec,
SSL/TLS, SRTP and MACsec up to 2000Mbps in multi-core SoCs
offering a large selection of cipher algorithms.

This patch only initializes the hardware without any interaction
with the Kernel crypto API, this allows to initialize the hardware
and use it by a user space driver which initializes the rings and
handles the runtime operation.

Change-Id: I66818fb761a298035fa9bf7b7442cb4c593dac2e
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37210
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>