crypto: inside-secure: register supported algorithms
This patch adds the registration of all supported algorithms by
current EIP197 driver:
	cipher: AES (CBC and ESB)
	hash: SHA1, SHA224, SSH256, HMAC-SHA1

Note that we register the algorithms only once, for the first engine.
Practically, currently, the Kernel uses only the first engine
and performs basic initialization of the second engine.
Working with 2 engines in the Kernel is planned as future work.

Change-Id: I201e78b77fa5db59194b5e23dbc826c6db630623
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37582
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>