fix: clk: mvebu: cp110: always enable gate clock 18 (eMMC & GOP)
Gate clock 18 feeds many core clocks in CP110, one of this clocks
is the eMMC clock. eMMC driver supports only one clock - the core
clock of eMMC) so we need to enable clock 18 in CP110 clock level
and not the eMMC driver itself.

This is a workaround, the complete solution should be nesting all
clock providers and consumers in the CP110 driver.

One possible issue with this patch might be the fact that if
we boot without IOs which use this clock, this clock will be
still enabled.

This patch prevents a stall when the Kernel is booted without
ppv2 driver (Clock 18 is shared between eMMC and GOP so currently,
eMMC "relys" on ppv2 for enabling this clock and this is a bug).

Change-Id: Iab49fd670f075c1c5d32b3edf30c6f27637ae117
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40500
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40641