irqchip: imx-gpcv2: Add explicit set affinity
Allow the affinity setting to be done also in GPC. On the i.MX8MQ SoC,
due to the ERR11171, the GIC is not able to wake up the cores and so,
by 'copying' the affinity to the GPC we allow the cores to wake up
when needed. One of these usecases is the interrupt of the system
counter which acts as broadcast device.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>