WIP: hack: arm64: dts: librem5-devkit: various buckets and regulators always-on
my devkit would stop with the following, exactly at the same time:

[   35.104345] PM: Adding info for No Bus:lo
[   35.327694] PM: Removing info for No Bus:lo
[   40.025566] 1v8_p: disabling
[   40.028468] 2v8_p: disabling
[   40.031406] buck1: disabling

I'm not sure why. Others don't see this happening. if only buck1 is set to
always-on, I get:

[   42.071568] buck2: disabling
[   42.075038] buck3: disabling

and with buck1, buck2 and buck3 always-on I still get stopped at

[   42.071604] buck5: disabling

and so on with buck6 ldo5

ldo6 ldo7 always on: usb xhci dies otherwise