Revert "rtc: m41t80: Clear oscillator fault bit during probe"
This reverts commit 206778f7a8412b3f694cee3d3931e4d77cd5af7b.

When the OF bit is set, we want to ignore the stored value
until it's overwritten by a new one because it's known to be
invalid. The OF bit is being cleared in m41t80_rtc_set_time
already, so once the kernel stores a valid value into RTC
it can get read on next boot without issues.

Closes #258